agents/amwal/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-dotenv
